fix a few misspellings
The pci.ids file ends up in golang projects.  Golang's linter is
quite picky and is commonly a continuous integration gating entity.
These misspelled words trip it.  It's fairly annoying then for
projects to need to special case and ignore errors from certain
files and not others.  Thus an attempt to get the spelling corrected
in the upstream source.

Signed-off-by: Tim Pepper <tpepper@vmware.com>